--- nbapi/endpoints/commonallplayers.py
@@ -1,9 +1,0 @@
-from nbapi.endpoints._base import Endpoint
-
-
-class CommonAllPlayers(Endpoint):
-    _endpoint = "commonallplayers"
-    _params = {
-        "LeagueID": "00",
-        "Season": "2020-21",
-    }
