--- nbapi/endpoints/commonplayerinfo.py
@@ -1,8 +1,0 @@
-from nbapi.endpoints._base import Endpoint
-
-
-class CommonPlayerInfo(Endpoint):
-    _endpoint = "commonplayerinfo"
-    _params = {
-        "PlayerID": None,
-    }
