+++ nbapi/endpoints/stats/commonteamyears.py
@@ -1,4 +1,4 @@
-from nbapi.endpoints._base import Endpoint
+from nbapi.endpoints.stats._base import Endpoint
 
 
 class CommonTeamYears(Endpoint):
