+++ nbapi/endpoints/stats/leaguedashplayerclutch.py
@@ -1,9 +1,11 @@
-from nbapi.endpoints._base import Endpoint
+from nbapi.endpoints.stats._base import Endpoint
 
 
-class LeagueDashPtStats(Endpoint):
-    _endpoint = "leaguedashptstats"
+class LeagueDashPlayerClutch(Endpoint):
+    _endpoint = "leaguedashplayerclutch"
     _params = {
+        "AheadBehind": "Ahead or Behind",
+        "ClutchTime": "Last 5 Minutes",
         "College": None,
         "Conference": None,
         "Country": None,
@@ -13,22 +15,28 @@ class LeagueDashPtStats(Endpoint):
         "DraftPick": None,
         "DraftYear": None,
         "GameScope": None,
+        "GameSegment": None,
         "Height": None,
         "LastNGames": 0,
         "LeagueID": "00",
         "Location": None,
+        "MeasureType": "Base",
         "Month": 0,
         "OpponentTeamID": 0,
         "Outcome": None,
         "PORound": 0,
+        "PaceAdjust": "N",
         "PerMode": "PerGame",
+        "Period": 0,
         "PlayerExperience": None,
-        "PlayerOrTeam": "Player",  # TODO: Player or Team here
         "PlayerPosition": None,
-        "PtMeasureType": "Drives",
+        "PlusMinus": "N",
+        "PointDiff": 5,
+        "Rank": "N",
         "Season": "2020-21",
         "SeasonSegment": None,
         "SeasonType": "Regular Season",
+        "ShotClockRange": None,
         "StarterBench": None,
         "TeamID": 0,
         "VsConference": None,
