+++ nbapi/endpoints/stats/leaguedashplayerptshot.py
@@ -1,9 +1,10 @@
-from nbapi.endpoints._base import Endpoint
+from nbapi.endpoints.stats._base import Endpoint
 
 
-class LeagueHustleStatsPlayer(Endpoint):
-    _endpoint = "leaguehustlestatsplayer"
+class LeagueDashPtShot(Endpoint):
+    _endpoint = "leaguedashptshot"
     _params = {
+        "CloseDefDistRange": None,
         "College": None,
         "Conference": None,
         "Country": None,
@@ -12,7 +13,11 @@ class LeagueHustleStatsPlayer(Endpoint):
         "Division": None,
         "DraftPick": None,
         "DraftYear": None,
+        "DribbleRange": None,
+        "GameSegment": None,
+        "GeneralRange": "Overall",
         "Height": None,
+        "LastNGames": 0,
         "LeagueID": "00",
         "Location": None,
         "Month": 0,
@@ -20,12 +25,17 @@ class LeagueHustleStatsPlayer(Endpoint):
         "Outcome": None,
         "PORound": 0,
         "PerMode": "PerGame",
+        "Period": 0,
         "PlayerExperience": None,
         "PlayerPosition": None,
         "Season": "2020-21",
         "SeasonSegment": None,
         "SeasonType": "Regular Season",
+        "ShotClockRange": None,
+        "ShotDistRange": None,
+        "StarterBench": None,
         "TeamID": 0,
+        "TouchTimeRange": None,
         "VsConference": None,
         "VsDivision": None,
         "Weight": None,
