+++ nbapi/endpoints/stats/leaguedashplayerstats.py
@@ -1,15 +1,14 @@
-from nbapi.endpoints._base import Endpoint
+from nbapi.endpoints.stats._base import Endpoint
 
 
-class LeagueDashPlayerShotLocations(Endpoint):
-    _endpoint = "leaguedashplayershotlocations"
+class LeagueDashPlayerStats(Endpoint):
+    _endpoint = "leaguedashplayerstats"
     _params = {
         "College": None,
         "Conference": None,
         "Country": None,
         "DateFrom": None,
         "DateTo": None,
-        "DistanceRange": "5ft Range",
         "Division": None,
         "DraftPick": None,
         "DraftYear": None,
@@ -37,6 +36,7 @@ class LeagueDashPlayerShotLocations(Endpoint):
         "ShotClockRange": None,
         "StarterBench": None,
         "TeamID": 0,
+        "TwoWay": 0,
         "VsConference": None,
         "VsDivision": None,
         "Weight": None,
