+++ nbapi/endpoints/stats/leaguedashptstats.py
@@ -1,37 +1,37 @@
-from nbapi.endpoints._base import Endpoint
+from nbapi.endpoints.stats._base import Endpoint
 
 
-class LeagueDashTeamStats(Endpoint):
-    _endpoint = "leaguedashteamstats"
+class LeagueDashPtStats(Endpoint):
+    _endpoint = "leaguedashptstats"
     _params = {
+        "College": None,
         "Conference": None,
+        "Country": None,
         "DateFrom": None,
         "DateTo": None,
         "Division": None,
+        "DraftPick": None,
+        "DraftYear": None,
         "GameScope": None,
-        "GameSegment": None,
+        "Height": None,
         "LastNGames": 0,
         "LeagueID": "00",
         "Location": None,
-        "MeasureType": "Base",
         "Month": 0,
         "OpponentTeamID": 0,
         "Outcome": None,
         "PORound": 0,
-        "PaceAdjust": "N",
         "PerMode": "PerGame",
-        "Period": 0,
         "PlayerExperience": None,
+        "PlayerOrTeam": "Player",  # TODO: Player or Team here
         "PlayerPosition": None,
-        "PlusMinus": "N",
-        "Rank": "N",
+        "PtMeasureType": "Drives",
         "Season": "2020-21",
         "SeasonSegment": None,
         "SeasonType": "Regular Season",
-        "ShotClockRange": None,
         "StarterBench": None,
         "TeamID": 0,
-        "TwoWay": 0,
         "VsConference": None,
         "VsDivision": None,
+        "Weight": None,
     }
