+++ nbapi/endpoints/stats/leaguedashteamptshot.py
@@ -1,37 +1,33 @@
-from nbapi.endpoints._base import Endpoint
+from nbapi.endpoints.stats._base import Endpoint
 
 
-class LeagueDashTeamShotLocations(Endpoint):
-    _endpoint = "leaguedashteamshotlocations"
+class LeagueDashTeamPtShot(Endpoint):
+    _endpoint = "leaguedashteamptshot"
     _params = {
+        "CloseDefDistRange": None,
         "Conference": None,
         "DateFrom": None,
         "DateTo": None,
-        "DistanceRange": "5ft Range",
         "Division": None,
-        "GameScope": None,
+        "DribbleRange": None,
         "GameSegment": None,
+        "GeneralRange": "Overall",
         "LastNGames": 0,
         "LeagueID": "00",
         "Location": None,
-        "MeasureType": "Base",
         "Month": 0,
         "OpponentTeamID": 0,
         "Outcome": None,
         "PORound": 0,
-        "PaceAdjust": "N",
         "PerMode": "PerGame",
         "Period": 0,
-        "PlayerExperience": None,
-        "PlayerPosition": None,
-        "PlusMinus": "N",
-        "Rank": "N",
         "Season": "2020-21",
         "SeasonSegment": None,
         "SeasonType": "Regular Season",
         "ShotClockRange": None,
-        "StarterBench": None,
+        "ShotDistRange": None,
         "TeamID": 0,
+        "TouchTimeRange": None,
         "VsConference": None,
         "VsDivision": None,
     }
