+++ nbapi/endpoints/stats/leaguedashteamstats.py
@@ -1,33 +1,37 @@
-from nbapi.endpoints._base import Endpoint
+from nbapi.endpoints.stats._base import Endpoint
 
 
-class LeagueDashTeamPtShot(Endpoint):
-    _endpoint = "leaguedashteamptshot"
+class LeagueDashTeamStats(Endpoint):
+    _endpoint = "leaguedashteamstats"
     _params = {
-        "CloseDefDistRange": None,
         "Conference": None,
         "DateFrom": None,
         "DateTo": None,
         "Division": None,
-        "DribbleRange": None,
+        "GameScope": None,
         "GameSegment": None,
-        "GeneralRange": "Overall",
         "LastNGames": 0,
         "LeagueID": "00",
         "Location": None,
+        "MeasureType": "Base",
         "Month": 0,
         "OpponentTeamID": 0,
         "Outcome": None,
         "PORound": 0,
+        "PaceAdjust": "N",
         "PerMode": "PerGame",
         "Period": 0,
+        "PlayerExperience": None,
+        "PlayerPosition": None,
+        "PlusMinus": "N",
+        "Rank": "N",
         "Season": "2020-21",
         "SeasonSegment": None,
         "SeasonType": "Regular Season",
         "ShotClockRange": None,
-        "ShotDistRange": None,
+        "StarterBench": None,
         "TeamID": 0,
-        "TouchTimeRange": None,
+        "TwoWay": 0,
         "VsConference": None,
         "VsDivision": None,
     }
