+++ nbapi/endpoints/stats/leaguehustlestatsplayer.py
@@ -1,18 +1,18 @@
-from nbapi.endpoints._base import Endpoint
+from nbapi.endpoints.stats._base import Endpoint
 
 
-class LeagueDashOppPtShot(Endpoint):
-    _endpoint = "leaguedashoppptshot"
+class LeagueHustleStatsPlayer(Endpoint):
+    _endpoint = "leaguehustlestatsplayer"
     _params = {
-        "CloseDefDistRange": None,
+        "College": None,
         "Conference": None,
+        "Country": None,
         "DateFrom": None,
         "DateTo": None,
         "Division": None,
-        "DribbleRange": None,
-        "GameSegment": None,
-        "GeneralRange": "Overall",
-        "LastNGames": 0,
+        "DraftPick": None,
+        "DraftYear": None,
+        "Height": None,
         "LeagueID": "00",
         "Location": None,
         "Month": 0,
@@ -20,14 +20,13 @@ class LeagueDashOppPtShot(Endpoint):
         "Outcome": None,
         "PORound": 0,
         "PerMode": "PerGame",
-        "Period": 0,
+        "PlayerExperience": None,
+        "PlayerPosition": None,
         "Season": "2020-21",
         "SeasonSegment": None,
         "SeasonType": "Regular Season",
-        "ShotClockRange": None,
-        "ShotDistRange": None,
         "TeamID": 0,
-        "TouchTimeRange": None,
         "VsConference": None,
         "VsDivision": None,
+        "Weight": None,
     }
