+++ nbapi/endpoints/stats/leaguehustlestatsteam.py
@@ -1,8 +1,8 @@
-from nbapi.endpoints._base import Endpoint
+from nbapi.endpoints.stats._base import Endpoint
 
 
-class LeagueDashPlayerBioStats(Endpoint):
-    _endpoint = "leaguedashplayerbiostats"
+class LeagueHustleStatsTeam(Endpoint):
+    _endpoint = "leaguehustlestatsteam"
     _params = {
         "College": None,
         "Conference": None,
@@ -12,10 +12,7 @@ class LeagueDashPlayerBioStats(Endpoint):
         "Division": None,
         "DraftPick": None,
         "DraftYear": None,
-        "GameScope": None,
-        "GameSegment": None,
         "Height": None,
-        "LastNGames": 0,
         "LeagueID": "00",
         "Location": None,
         "Month": 0,
@@ -23,14 +20,11 @@ class LeagueDashPlayerBioStats(Endpoint):
         "Outcome": None,
         "PORound": 0,
         "PerMode": "PerGame",
-        "Period": 0,
         "PlayerExperience": None,
         "PlayerPosition": None,
         "Season": "2020-21",
         "SeasonSegment": None,
         "SeasonType": "Regular Season",
-        "ShotClockRange": None,
-        "StarterBench": None,
         "TeamID": 0,
         "VsConference": None,
         "VsDivision": None,
