+++ nbapi/endpoints/stats/synergyplaytypes.py
@@ -1,4 +1,4 @@
-from nbapi.endpoints._base import Endpoint
+from nbapi.endpoints.stats._base import Endpoint
 
 
 class SynergyPlayTypes(Endpoint):
